+++ crates/runtime/src/lib.rs
@@ -29,7 +29,7 @@ use ink_revive_types::{
 pub use macros::{
     AssetIdForTrustBackedAssets,
     BlockBuilder,
-    DefaultSandbox,
+    DefaultRuntime,
     TrustBackedAssetsInstance,
 };
 use pallet_revive::{
@@ -38,7 +38,7 @@ use pallet_revive::{
     InstantiateReturnValue,
 };
 use sp_core::Get;
-/// Export pallets that are used in [`crate::create_sandbox`]
+/// Export pallets that are used in [`crate::create_runtime`]
 pub use {
     frame_support::sp_runtime::testing::H256,
     frame_support::{
@@ -67,7 +67,7 @@ pub use {
 };
 
 pub use client::{
-    Client as SandboxClient,
+    Client as RuntimeClient,
     preset,
 };
 pub use error::E2EError;
@@ -117,15 +117,15 @@ pub type ContractExecResultFor<Runtime> =
 /// Alias for the `map_account` result.
 pub type MapAccountResultFor = Result<(), DispatchError>;
 
-/// Alias for the runtime of a sandbox.
-pub type RuntimeOf<S> = <S as Sandbox>::Runtime;
+/// Alias for the runtime of a runtime environment.
+pub type RuntimeOf<S> = <S as RuntimeEnv>::Runtime;
 
-/// Alias for the runtime event of a sandbox.
+/// Alias for the runtime event of a runtime environment.
 pub type RuntimeEventOf<S> = <RuntimeOf<S> as frame_system::Config>::RuntimeEvent;
 
-/// Sandbox defines the API of a sandboxed runtime.
-pub trait Sandbox {
-    /// The runtime associated with the sandbox.
+/// RuntimeEnv defines the API of a runtime environment.
+pub trait RuntimeEnv {
+    /// The runtime associated with the runtime environment.
     type Runtime: frame_system::Config;
 
     /// Execute the given externalities.
@@ -151,7 +151,7 @@ pub trait Sandbox {
         Default::default()
     }
 
-    /// Default actor for the sandbox.
+    /// Default actor for the runtime environment.
     fn default_actor() -> AccountIdFor<Self::Runtime>;
 
     fn default_gas_limit() -> Weight {
@@ -194,7 +194,7 @@ where
     native_to_eth_ratio.saturating_mul(evm_value)
 }
 
-/// Convert a `pallet_revive::CallTrace` (sandbox) into an `ink_revive_types::CallTrace`
+/// Convert a `pallet_revive::CallTrace` (runtime) into an `ink_revive_types::CallTrace`
 /// (API).
 pub fn to_revive_trace(t: pallet_revive::evm::CallTrace) -> CallTrace {
     CallTrace {
@@ -258,7 +258,7 @@ pub fn to_revive_storage_deposit<B>(
 
 /// Trait for types that can be converted into a runtime AccountId.
 ///
-/// This allows sandbox APIs to accept various account types without requiring manual
+/// This allows runtime APIs to accept various account types without requiring manual
 /// conversion.
 pub trait IntoAccountId<AccountId> {
     fn into_account_id(self) -> AccountId;