+++ integration-tests/public/runtime-call-contract/custom-runtime/pallet-revive-caller/src/executor.rs
@@ -60,7 +60,7 @@ where
         let result = pallet_revive::Pallet::<R>::bare_call(
             self.origin.clone(),
             self.contract,
-            ink_sandbox::balance_to_evm_value::<R>(self.value),
+            ink_runtime::balance_to_evm_value::<R>(self.value),
             self.gas_limit,
             // self.storage_deposit_limit,
             BalanceOf::<R>::max_value(), // todo